test: Position 자료형 Int -> Long 수정

+++ src/main/kotlin/racingcar/Model/Car.kt
@@ -3,9 +3,9 @@ package racingcar.Model
 import camp.nextstep.edu.missionutils.Randoms
 
 class Car(val name : String) {
-    private var position = 0
+    private var position = 0L
     fun goOrStop() { position += goOrStopDetermine() }
-    fun getPosition() : Int = position
+    fun getPosition() : Long = position
 
     //outputview로 이동 예정
     fun getCurrentPosition() = "$name : " + lengthToPrint(position)
@@ -13,7 +13,7 @@ class Car(val name : String) {
     //For test Code
     fun addPosition() { position += 1 }
 
-    private fun lengthToPrint(length : Int) : String {
+    private fun lengthToPrint(length : Long) : String {
         var position = ""
         for(i in 0 until length) position += "-"
         return position
